Release/beta4.1

+++ ClassroomManagerAPI/Application/Commands/Classroom/DeleteClassroomCommand.cs
@@ -1,8 +1,9 @@
 ﻿using AutoMapper;
 using ClassroomManagerAPI.Common;
-using ClassroomManagerAPI.Enums;
+using ClassroomManagerAPI.Enums.ErrorCodes;
 using ClassroomManagerAPI.Repositories.IRepositories;
 using MediatR;
+using Microsoft.EntityFrameworkCore;
 using System.Net;
 
 namespace ClassroomManagerAPI.Application.Commands.Classroom
@@ -24,6 +25,13 @@ namespace ClassroomManagerAPI.Application.Commands.Classroom
 		{
 			ArgumentNullException.ThrowIfNull(request);
 			ResponseMethod<string> result = new ResponseMethod<string>();
+			var classroom = _classroomRepository.Queryable.Include(x => x.Schedules).Any(x => (!x.IsDeleted) && (x.FacilityAmount != 0 || x.Schedules.Any(f => f.EndTime <= DateTime.Now.AddDays(7))));
+			if (classroom)
+			{
+				result.AddBadRequest(nameof(ErrorClassEnum.ConflictSchedule));
+				result.StatusCode = (int)HttpStatusCode.Conflict;
+				return result;
+			}
 			var deletedClassrooom = await _classroomRepository.DeleteAsync(request.Id).ConfigureAwait(false);
 			if (!deletedClassrooom)
 			{
